src/requirements.txt
elasticsearch
mysql-connector-python
requests
tqdm
